refactor(BA-6919): retype shared action sinks to the new EntityType
Point ActionMetricObserver.observe_action and the reporter
Started/FinishedActionMessage dataclasses at
common.entity.types.EntityType (a str NewType) instead of the legacy
RBAC common.data.permission.types.EntityType StrEnum. The new
self-contained base actions already return the new EntityType, so this
removes the type mismatch at the monitor boundary without any cast.

Legacy BaseAction-derived monitors still yield the RBAC StrEnum; wrap
it with EntityType(...) at the prometheus/reporter monitor boundary,
which is well-typed because the StrEnum is a str subtype. This keeps
the legacy monitor path compiling.
