+++ src/ai/backend/manager/actions/monitors/reporter.py
@@ -2,6 +2,7 @@ from typing import override
 
 from ai.backend.common.contexts.request_id import current_request_id
 from ai.backend.common.contexts.user import current_user, triggered_user
+from ai.backend.common.entity.types import EntityType
 from ai.backend.manager.actions.action import BaseAction, BaseActionTriggerMeta, ProcessResult
 from ai.backend.manager.actions.monitors.monitor import ActionMonitor
 from ai.backend.manager.actions.types import BLANK_ID
@@ -25,7 +26,7 @@ class ReporterMonitor(ActionMonitor):
             action_id=meta.action_id,
             action_type=action.spec().type(),
             entity_id=action.entity_id(),
-            entity_type=action.entity_type(),
+            entity_type=EntityType(action.entity_type()),
             request_id=current_request_id(),
             triggered_by=str(trigger.user_id) if trigger else None,
             acted_as=acting.user_id if acting else None,
@@ -45,7 +46,7 @@ class ReporterMonitor(ActionMonitor):
             request_id=current_request_id() or BLANK_ID,
             triggered_by=str(trigger.user_id) if trigger else None,
             acted_as=acting.user_id if acting else None,
-            entity_type=action.entity_type(),
+            entity_type=EntityType(action.entity_type()),
             operation_type=action.operation_type(),
             status=result.meta.status,
             description=result.meta.description,
